added tests, enum bassNote, exception handling

+++ src/main/java/com/karaoke/karaokemaker/controllers/ChordController.java
@@ -1,11 +1,13 @@
 package com.karaoke.karaokemaker.controllers;
 
+import com.karaoke.karaokemaker.exceptions.ResourceNotFoundException;
 import com.karaoke.karaokemaker.model.Chord;
 import com.karaoke.karaokemaker.repositories.ChordRepository;
 import com.karaoke.karaokemaker.service.ChordService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
+
 import java.util.List;
 
 
@@ -22,37 +24,45 @@ public class ChordController {
 
     @PostMapping("/add")
     @PreAuthorize("hasRole('ROLE_ADMIN')")
-    public ResponseEntity<Chord>  postChord(@RequestBody Chord chord) {
-        return ResponseEntity.ok (chordService.add(chord));
+    public ResponseEntity<Chord> postChord(@RequestBody Chord chord) {
+        return ResponseEntity.ok(chordService.saveChord(chord));
     }
 
     @GetMapping("/all")
-    public ResponseEntity<List<Chord>>  getChords() {
-        return ResponseEntity.ok (chordService.getChords());
+    public ResponseEntity<List<Chord>> getChords() {
+        return ResponseEntity.ok(chordService.getAllChords());
     }
 
     @GetMapping("/{id}")
     public ResponseEntity<?> getSignleChord(@PathVariable Long id) {
 
-        return  chordService.getSingleChord(id)
-                .map(c -> ResponseEntity.ok(c))
-                .orElse(ResponseEntity.notFound().build());
+        Chord chord = chordService.getSingleChord(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Chord with ID :" + id + " Not Found"));
+
+        return ResponseEntity.ok().body(chord);
+
     }
+
 
     @DeleteMapping("/{id}")
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     ResponseEntity<?> deleteChord(@PathVariable Long id) {
+
+        if (chordRepository.findById(id).isEmpty()) {
+            throw new ResourceNotFoundException("Chord with ID :" + id + " Not Found");
+        }
         chordService.deleteChord(id);
-        System.out.println("chord deleted");
         return ResponseEntity.noContent().build();
     }
 
     @PutMapping("/{id}")
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     ResponseEntity<?> putChord(@PathVariable Long id, @RequestBody Chord chord) {
-        return chordService.replaceChord(id, chord)
-                .map(c -> ResponseEntity.noContent().build())
-                .orElse(ResponseEntity.notFound().build());
+        Chord updatedChord = chordService.replaceChord(id, chord)
+                .orElseThrow(() -> new ResourceNotFoundException("Chord with ID :" + id + " Not Found"));
+
+
+        return ResponseEntity.noContent().build();
     }
 
 }
